fix(compiler): `/` is floating-point division, on every backend
Chip task_dacd709d: `7 / 2` evaluated to 3 on BOTH the C# and C++ backends.
In VB.NET `/` is ALWAYS floating point -- 7 / 2 is 3.5, typed Double -- and `\`
is the integer operator. BasicLang typed Integer / Integer as Integer, so both
backends inherited C-family truncation. Neither backend was at fault; the JS
backend was already correct and JavaScriptTypeMapper even carries a comment
warning about the hazard below.

FOUR SITES, ALL LOAD-BEARING, MUTATION-VERIFIED

  1. SemanticAnalyzer: `/` no longer shares the `+ - * % Mod` arm's GetCommonType.
     Decimal stays Decimal and Single stays Single -- promoting Decimal would lose
     precision and it has no implicit conversion to Double in either direction.
  2. IROptimizer.FoldDiv returns a double for the int/int and long/long arms.
     Independent of 1, proven via the JS backend: in ONE program the variable form
     printed 3.5 while the constant form printed 3. Must ship WITH 1 because
     TryFoldBinary stamps op.Type onto the folded constant -- alone, 1 yields
     IRConstant(3 as int) typed Double and 2 yields `int x = 3.5`, which will not
     compile.
  3. The `\` arm now accepts a floating operand (VB converts it to Long first).
     REQUIRED BY 1, not cleanup: `(a / b) \ c` compiles today only because `a / b`
     is typed Integer.
  4. IRBuilder widens the OPERANDS of a floating division via IRCast.

SITE 4 IS THE ONE EMISSION TESTS COULD NOT SEE

Neither backend reads IRBinaryOp.Type. Both render `{left} {op} {right}` and let
the TARGET language choose the semantics. With only sites 1-3, the emitted C++ was

    double t0 = {};
    t0 = a / b;      // int32_t / int32_t -- truncates, THEN widens

The test asserted the temp was `double` and passed. The compiled binary printed 3.
Only running it caught this, which is exactly why this repo's rule says stdout is
the oracle and emission assertions are third-best. The test now asserts
`static_cast<double>(a)` on the OPERANDS, and its comment records why.

IRCast was chosen over per-backend coercion because it is already handled by both
interpreters, all five backends, IROperandWalker, IRPrettyPrinter and
CppCapabilityChecker -- one insertion moves every consumer instead of repeating
the fix per backend and missing LLVM/MSIL/JS.

NARROWING IS NOW IMPLICIT, AND THAT IS A DELIBERATE LANGUAGE CALL

Correct `/` makes `Dim x As Integer = a / b` a narrowing conversion. Rejecting it
is what VB.NET does under Option Strict On -- but BasicLang has NO Option Strict.
Verified rather than assumed: `Option Public` is the only Option form the language
implements (Parser.cs, ModuleResolver.cs, Compiler.cs), and `Option Explicit` is a
hard parse error despite the LSP offering a quick-fix that inserts it (chipped
separately). With no way to opt out, a strict rule would simply break existing
source.

So TypeInfo.IsAssignableFrom now permits floating-to-integral and integral-to-
integral narrowing. The conversion lowers the way the explicit CInt() already
does -- truncation, not VB's banker's rounding -- so existing programs keep
compiling AND keep their current values. `Dim x As Integer = 7 / 2` is still 3.
Only the value never stored into an integer changes, which is the reported bug.

The integral-to-integral half is not optional either: with 3 in place,
`(a / b) \ c` is Long, so returning it from an Integer function is a narrowing.
Permitting only the floating case would leave the language half-strict and still
regress working source.

  D1  operand widening disabled  -> 2 red (both division-value tests)
  D2  FoldDiv reverted           -> covered by the pre-fix red run
  D3  analyzer promotion removed -> covered by the pre-fix red run
  D4  narrowing permission off   -> 2 red (both compatibility tests)

ONE TRADE ACCEPTED

`7 / 2` no longer folds to a literal, because widening wraps both constants in
casts the folder does not see through. Correctness is unaffected -- the value is
3.5 either way, verified by running the binary -- and g++/clang and the JIT fold
it downstream. Teaching IROptimizer to fold cast-of-constant would recover the
literal, but casts carry narrowing semantics and that is an optimizer feature,
not part of this fix.

VALUE PROVEN

Built the generated C++ with g++ 14.2 -std=c++20 and ran it:
7 / 2 -> 3.5 | a / b -> 3.5 | 7 \ 2 -> 3 | a \ b -> 3 | 1 / 3 -> 0.333333 |
-7 / 2 -> -3.5 | Dim x As Integer = 7 / 2 -> 3.

Gates: fast subset 4583 passed, 2 pre-existing SearchSnippets failures
(task_b9620d48), 1 skipped -- prior 4576 plus exactly the 7 tests added here.

+++ BasicLang/IROptimizer.cs
@@ -464,10 +464,18 @@ namespace BasicLang.Compiler.IR.Optimization
             return null;
         }
         
+        // `/` is FLOATING-POINT division in VB.NET even on two integers: 7 / 2 is 3.5.
+        // These two arms used C# integer division, so the CONSTANT path truncated
+        // independently of the analyzer's result type — proven by the JS backend, which is
+        // otherwise correct yet still printed 3 for `7 / 2` while printing 3.5 for `a / b`.
+        // Must ship with the analyzer's Double result type: TryFoldBinary stamps op.Type onto
+        // the folded IRConstant, so fixing either one alone yields a mismatched pair
+        // (IRConstant(3 as int) typed Double, or IRConstant(3.5) typed Integer).
+        // FoldIntDiv below is deliberately left truncating — `\` is the integer operator.
         private object FoldDiv(object a, object b)
         {
-            if (a is int ia && b is int ib && ib != 0) return ia / ib;
-            if (a is long la && b is long lb && lb != 0) return la / lb;
+            if (a is int ia && b is int ib && ib != 0) return (double)ia / ib;
+            if (a is long la && b is long lb && lb != 0) return (double)la / lb;
             if (a is float fa && b is float fb && fb != 0) return fa / fb;
             if (a is double da && b is double db && db != 0) return da / db;
             return null;
